requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask[async]==3.0.0
Flask-APScheduler==1.13.1
APScheduler==3.10.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
Werkzeug==3.0.1
yq==3.2.3
zipp==3.17.0
requests==2.31.0
sqlalchemy==2.0.23
flaskcode==0.0.8
pyyaml==5.4.1